usr.bin/cc: split value into it's own parts
