+++ usr.bin/cc/Tupfile
@@ -3,8 +3,9 @@ include_rules
 # toolchain build
 : lexer.l |> !host_lex |>
 
-: foreach parser/*.c  |> !host_cc   |> ./obj/parser/%B.o {obj}
-: foreach *.c ^test.c |> !host_cc   |> ./obj/%B.o {obj}
-: {obj}               |> !host_ccld |> $(ROOT)/toolchain/cc | $(ROOT)/<toolchain>
+: foreach parser/value/*.c |> !host_cc   |> ./obj/parser/value/%B.o {obj}
+: foreach parser/*.c       |> !host_cc   |> ./obj/parser/%B.o {obj}
+: foreach *.c ^test.c      |> !host_cc   |> ./obj/%B.o {obj}
+: {obj}                    |> !host_ccld |> $(ROOT)/toolchain/cc | $(ROOT)/<toolchain>
 
 : test.c | $(ROOT)/<toolchain> |> $(VALGRIND) $(ROOT)/toolchain/cc < test.c |>
